requirements.txt
unsloth
transformers>=4.51.0
datasets>=3.0.0
trl>=0.24.0
accelerate>=0.34.0
peft>=0.13.0
bitsandbytes>=0.44.0
pyyaml>=6.0.2
jsonschema>=4.23.0
tqdm>=4.66.0
openai>=1.60.0
pandas>=2.0.0
huggingface_hub>=0.24.0
